+++ scripts/macros.js
@@ -290,6 +290,7 @@ export {preciseHunter} from './macros/2024/classFeatures/ranger/preciseHunter.js
 export {foeSlayer} from './macros/2024/classFeatures/ranger/foeSlayer.js';
 //Beast Master
 export {primalCompanion} from './macros/2024/classFeatures/ranger/beastMaster/primalCompanion.js';
+export {exceptionalTraining} from './macros/2024/classFeatures/ranger/beastMaster/exceptionalTraining.js';
 //Rogue
 export {sneakAttack} from './macros/2024/classFeatures/rogue/sneakAttack.js';
 export {cunningAction} from './macros/2024/classFeatures/rogue/cunningAction.js';